backend/utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Shared utilities for microservices."
dependencies = [
    "pyjwt[crypto]",
    "redis>=5.0.0",
    "fastapi",
]
requires-python = ">=3.11"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
utils = "."

[tool.setuptools]
packages = ["utils"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv]
